remove old update checker and stale code
