+++ nuget/lib/dependabot/nuget/discovery/workspace_discovery.rb
@@ -2,8 +2,8 @@
 # frozen_string_literal: true
 
 require "dependabot/nuget/discovery/dependency_file_discovery"
-require "dependabot/nuget/discovery/directory_packages_props_discovery"
 require "dependabot/nuget/discovery/project_discovery"
+require "dependabot/nuget/native_helpers"
 require "sorbet-runtime"
 
 module Dependabot
@@ -13,48 +13,43 @@ module Dependabot
 
       sig { params(json: T::Hash[String, T.untyped]).returns(WorkspaceDiscovery) }
       def self.from_json(json)
-        file_path = T.let(json.fetch("FilePath"), String)
-        projects = T.let(json.fetch("Projects"), T::Array[T::Hash[String, T.untyped]]).filter_map do |project|
-          ProjectDiscovery.from_json(project)
-        end
-        directory_packages_props = DirectoryPackagesPropsDiscovery
-                                   .from_json(T.let(json.fetch("DirectoryPackagesProps"),
-                                                    T.nilable(T::Hash[String, T.untyped])))
-        global_json = DependencyFileDiscovery
-                      .from_json(T.let(json.fetch("GlobalJson"), T.nilable(T::Hash[String, T.untyped])))
-        dotnet_tools_json = DependencyFileDiscovery
-                            .from_json(T.let(json.fetch("DotNetToolsJson"), T.nilable(T::Hash[String, T.untyped])))
+        Dependabot::Nuget::NativeHelpers.ensure_no_errors(json)
 
-        WorkspaceDiscovery.new(file_path: file_path,
+        path = T.let(json.fetch("Path"), String)
+        path = "/" + path unless path.start_with?("/")
+        projects = T.let(json.fetch("Projects"), T::Array[T::Hash[String, T.untyped]]).filter_map do |project|
+          ProjectDiscovery.from_json(project, path)
+        end
+        global_json = DependencyFileDiscovery
+                      .from_json(T.let(json.fetch("GlobalJson"), T.nilable(T::Hash[String, T.untyped])), path)
+        dotnet_tools_json = DependencyFileDiscovery
+                            .from_json(T.let(json.fetch("DotNetToolsJson"),
+                                             T.nilable(T::Hash[String, T.untyped])), path)
+
+        WorkspaceDiscovery.new(path: path,
                                projects: projects,
-                               directory_packages_props: directory_packages_props,
                                global_json: global_json,
                                dotnet_tools_json: dotnet_tools_json)
       end
 
       sig do
-        params(file_path: String,
+        params(path: String,
                projects: T::Array[ProjectDiscovery],
-               directory_packages_props: T.nilable(DirectoryPackagesPropsDiscovery),
                global_json: T.nilable(DependencyFileDiscovery),
                dotnet_tools_json: T.nilable(DependencyFileDiscovery)).void
       end
-      def initialize(file_path:, projects:, directory_packages_props:, global_json:, dotnet_tools_json:)
-        @file_path = file_path
+      def initialize(path:, projects:, global_json:, dotnet_tools_json:)
+        @path = path
         @projects = projects
-        @directory_packages_props = directory_packages_props
         @global_json = global_json
         @dotnet_tools_json = dotnet_tools_json
       end
 
       sig { returns(String) }
-      attr_reader :file_path
+      attr_reader :path
 
       sig { returns(T::Array[ProjectDiscovery]) }
       attr_reader :projects
-
-      sig { returns(T.nilable(DirectoryPackagesPropsDiscovery)) }
-      attr_reader :directory_packages_props
 
       sig { returns(T.nilable(DependencyFileDiscovery)) }
       attr_reader :global_json
